Lint Markdown files
A new Earthly target that lints Markdown files has been created. The
target executes a function that can be used by other projects as well.
The function installs the `markdownlint-cli` from npm, copies all files
in the working directory into the container, and then runs the CLI on
all Markdown files.

The version of `markdownlint-cli` can be passed as an argument to the
function. In this repository, we are happy to use the latest version,
but other projects might want to lock the version.