Lint YAML files
A new Earthly target that lints YAML files has been created. The target
executes a function that can be used by other projects as well. The
function pulls a Docker image with `yamllint`, copies all files in the
directory into the container, and then runs `yamllint` on all YAML
files.

The version of the `yamllint` image can be passed as an argument to the
function. In this repository, we are happy to use the latest version,
but other projects might want to lock the version.